dataTable 转model 工具类编写

+++ MyAspNhOracle/Models/Owners.cs
@@ -8,13 +8,13 @@ namespace MyAspNhOracle.Models
 {
     public class Owners
     {
-        public virtual int Id { get; set; }
+        public virtual decimal Id { get; set; }
         public virtual string Name { get; set; }
-        public virtual int AddressId { get; set; }
+        public virtual decimal AddressId { get; set; }
         public virtual string HouseNumber { get; set; }
         public virtual string WaterMeter { get; set; }
         public virtual DateTime AddDate { get; set; }
-        public virtual int OwnerTypeId { get; set; }
+        public virtual decimal OwnerTypeId { get; set; }
 
 
     }
